+++ src/vumi2/applications/junebug_message_api/junebug_message_api.py
@@ -7,7 +7,7 @@ from httpx import AsyncClient
 from quart import request
 from trio import move_on_after
 
-from vumi2.cli import class_from_string
+from vumi2.class_helpers import class_from_string
 from vumi2.messages import (
     Event,
     Message,
@@ -105,6 +105,7 @@ class JunebugMessageApi(BaseWorker):
     config: JunebugMessageApiConfig
 
     async def setup(self) -> None:
+        await super().setup()
         state_cache_class = class_from_string(self.config.state_cache_class)
         self.state_cache = state_cache_class(self.config.state_cache_config)
         self.connector = await self.setup_receive_inbound_connector(