+++ src/vumi2/transports/smpp/smpp.py
@@ -4,7 +4,7 @@ from async_amqp import AmqpProtocol  # type: ignore
 from attrs import Factory, define
 from trio import MemoryReceiveChannel, Nursery, open_memory_channel, open_tcp_stream
 
-from vumi2.cli import class_from_string
+from vumi2.class_helpers import class_from_string
 from vumi2.messages import Event, Message, MessageType
 from vumi2.workers import BaseConfig, BaseWorker
 
@@ -69,6 +69,7 @@ class SmppTransceiverTransport(BaseWorker):
     async def setup(self) -> None:
         # We open the TCP connection first, so that we have a place to send any
         # outbounds once we start receiving them from the AMQP server
+        await super().setup()
         self.stream = await open_tcp_stream(
             host=self.config.host, port=self.config.port
         )