+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Retry messages if Turn rate limits us
 - Warning logs for config errors, including (ignored) extra keys
 - Log an error if the outbound request times out
+- Adding middleware support 
+- Adding logging and unidecoder middleware
+- Calling BaseWorker setup on all child classes
 
 ### Fixed
 - Sphinx builds now fail on warnings, and all existing warnings/errors in the docs have been fixed